+++ CHANGELOG.md
@@ -5,6 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+ - List
+   - fmap (map)
+   - operator[]
+   - mempty and mappend
+   - liftA2
+   - associative binary operation for Alternative concept
+   - filter
+   - foldl
 
 ## [0.3.0] - 2020-11-09
 ### Added